requirements.txt
tensorflow
numpy
pillow
scikit-learn
streamlit
tqdm
matplotlib
